examples/sse/requirements.txt
psutil==1.2.1
pyzmq==14.0.1